+++ Data-Structures-Project1/Data-Structures-Project1/SelectionInsertionSortAlgorithm.h
@@ -1,4 +1,4 @@
 #pragma once
 
-int selectionProblemUsingInsertionSort(int* arr, int arrLen, int index);
-void insertionSortAlgorithm(int* arr, int arrLen);
+double selectionProblemUsingInsertionSort(double* arr, int arrLen, int index);
+void insertionSortAlgorithm(double* arr, int arrLen);
